+++ lib/watir/locators/text_field/locator.rb
@@ -32,9 +32,6 @@ module Watir
           element = super
 
           if element && !Watir::TextField::NON_TEXT_TYPES.include?(element.attribute(:type))
-            if element.tag_name.downcase == 'textarea'
-              warn "Locating textareas with '#text_field' is deprecated. Please, use '#textarea' method instead."
-            end
             element
           end
         end